Python/3rdParty/Faust/Learn/01-hello-world/pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Antonio Alvarado Hernández", email = "tnotstar@gmail.com"},
]
dependencies = [
    "faust-streaming>=0.10.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
